Fix display of unfurled Slack links
For now, we cannot handle double quote characters in descriptions, since they
break the HTML and tera does not give us a smooth way to remove them
through filters.

Also, add additional opengraph tags. The side effect of this is that a blog
article must have a corresponding image, since tera does not allow you to check
for the existence/absence of a file through functions.