Fix Bugs with RSS & Responsive Tag / Category Listing Pages
RSS configuration and template logic changed in a new version of Zola. Update
the base.html and config.toml files accordingly.

A typo in style.scss caused the listing pages for tags and categories to not
properly take up the whole width of the screen. Fixing the typo resolves this.